refactor(moderation): update JailStatusCache usage for async operations
- Modified JailStatusCache calls in the jail and unjail modules to use async methods for cache invalidation, ensuring proper handling of asynchronous operations.
- Cleaned up import statements to consistently use the new cache import path across modules.
- Updated validation checks in utility functions for improved clarity and type safety.
