+++ src/tux/modules/moderation/unjail.py
@@ -11,12 +11,12 @@ import discord
 from discord.ext import commands
 from loguru import logger
 
+from tux.cache import JailStatusCache
 from tux.core.bot import Tux
 from tux.core.checks import requires_command_permission
 from tux.core.flags import UnjailFlags
 from tux.database.models import Case
 from tux.database.models import CaseType as DBCaseType
-from tux.shared.cache import JailStatusCache
 
 from . import ModerationCogBase
 
@@ -250,7 +250,7 @@ class Unjail(ModerationCogBase):
 
             # Invalidate jail status cache after unjailing
             assert ctx.guild is not None
-            JailStatusCache().invalidate(ctx.guild.id, member.id)
+            await JailStatusCache().invalidate(ctx.guild.id, member.id)
 
         # Execute the action (removed lock since moderation service handles concurrency)
         await perform_unjail()
